🎁🎉 Major Update v2.0.0
- `aria2` as download method.
- Cancel button for cancelling download process (`--gradio-queue` required)
- Debug snapshot.<br/>
When `take_snapshot()` is uncommented, it saves the current state of the webui on various location (into `snapshot.txt`), and when you type `#debugdebugdebug` on the textbox (and nothing more), it will compare the current state and the last saved state, and removes every new file/folder. This will be useful for debugging and testing.
- New hashtags: `#textualinversion`, `#ti`, `#aestheticembedding`, `#aestheticembed`, `#controlnet`, and `#cnet`
- `shlex.quote` to properly quote links (Thanks **[@rti7743](https://github.com/rti7743)**!)
- Supports cloning webui extensions
- Supports download from CivitAI model links (Thanks **[@rti7743](https://github.com/rti7743)**!)
- Supports for aesthetic gradients, controlnet model, and extensions path.
- Uses `subprocess.Popen` instead of `os.system`

Co-Authored-By: rti <297663+rti7743@users.noreply.github.com>